+++ mail/mailer_test.go
@@ -21,6 +21,7 @@ import (
 func TestSender(t *testing.T) {
 	t.Run("create auth if username and password are set", func(t *testing.T) {
 		config := configuration.Smtp{
+			Enabled:  true,
 			Username: "user",
 			Password: "pw",
 		}
@@ -34,7 +35,9 @@ func TestSender(t *testing.T) {
 	})
 
 	t.Run("return nil auth if empty", func(t *testing.T) {
-		config := configuration.Smtp{}
+		config := configuration.Smtp{
+			Enabled: true,
+		}
 		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
 
 		sender := CreateSender(config, "source", []string{}, mGlobalConfigRepo)
@@ -45,8 +48,9 @@ func TestSender(t *testing.T) {
 
 	t.Run("will create server address", func(t *testing.T) {
 		config := configuration.Smtp{
-			Port:   123,
-			Server: "server",
+			Enabled: true,
+			Port:    123,
+			Server:  "server",
 		}
 		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
 
@@ -65,8 +69,29 @@ func TestSender(t *testing.T) {
 }
 
 func TestSendMigrationResult(t *testing.T) {
+	t.Run("don't send a result email if email dispatch is disabled", func(t *testing.T) {
+		config := configuration.Smtp{
+			Enabled: false,
+		}
+
+		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
+
+		sender := CreateSender(config, "source", []string{}, mGlobalConfigRepo)
+		senderFunc := NewMockSenderService(t)
+		sender.senderService = func(addr string, a smtp.Auth, from string, to []string, msg []byte) error {
+			return senderFunc.Execute(addr, a, from, to, msg)
+		}
+
+		timeA, _ := time.Parse("15:04", "13:01")
+		timeB := timeA.Add(5 * time.Minute)
+
+		err := sender.Send(context.Background(), true, nil, timeA, timeB)
+		require.NoError(t, err)
+		senderFunc.AssertNotCalled(t, "Execute")
+	})
 	t.Run("send migration result", func(t *testing.T) {
 		config := configuration.Smtp{
+			Enabled:  true,
 			Server:   "server",
 			Port:     25,
 			Username: "username",
@@ -112,9 +137,10 @@ func TestSendMigrationResult(t *testing.T) {
 
 	t.Run("should send migration result with unauthenticated server", func(t *testing.T) {
 		config := configuration.Smtp{
-			Server: "server",
-			Port:   25,
-			From:   "from",
+			Enabled: true,
+			Server:  "server",
+			Port:    25,
+			From:    "from",
 			To: []string{
 				"a@test.de",
 				"b@test.de",
@@ -148,6 +174,7 @@ func TestSendMigrationResult(t *testing.T) {
 
 	t.Run("should fail to send migration result for error getting target-instance", func(t *testing.T) {
 		config := configuration.Smtp{
+			Enabled:  true,
 			Server:   "server",
 			Port:     25,
 			Username: "username",
@@ -188,6 +215,7 @@ func TestSendMigrationResult(t *testing.T) {
 
 	t.Run("should fail to send migration result for error getting target-instance while writing body-text", func(t *testing.T) {
 		config := configuration.Smtp{
+			Enabled:  true,
 			Server:   "server",
 			Port:     25,
 			Username: "username",
@@ -236,6 +264,7 @@ func TestSendMigrationResult(t *testing.T) {
 
 	t.Run("should not fail to send migration result for error reading attachment-file", func(t *testing.T) {
 		config := configuration.Smtp{
+			Enabled:  true,
 			Server:   "server",
 			Port:     25,
 			Username: "username",
@@ -284,7 +313,8 @@ func TestSendMigrationResult(t *testing.T) {
 		}()
 
 		config := configuration.Smtp{
-			Port: 25,
+			Enabled: true,
+			Port:    25,
 		}
 		sender := CreateSender(config, "source", []string{"a", "b"}, nil)
 
@@ -307,7 +337,8 @@ func TestSendMigrationResult(t *testing.T) {
 		}()
 
 		config := configuration.Smtp{
-			Server: "myserver",
+			Enabled: true,
+			Server:  "myserver",
 		}
 		sender := CreateSender(config, "source", []string{"a", "b"}, nil)
 
@@ -385,7 +416,9 @@ func TestSender_writeBodyText(t *testing.T) {
 	testCtx := context.Background()
 
 	t.Run("write body text for successful final migration", func(t *testing.T) {
-		config := configuration.Smtp{}
+		config := configuration.Smtp{
+			Enabled: true,
+		}
 		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
 		mGlobalConfigRepo.EXPECT().Get(testCtx).Return(
 			regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{
@@ -409,7 +442,9 @@ func TestSender_writeBodyText(t *testing.T) {
 	})
 
 	t.Run("write body text for delta migration with error", func(t *testing.T) {
-		config := configuration.Smtp{}
+		config := configuration.Smtp{
+			Enabled: true,
+		}
 		mGlobalConfigRepo := newMockGlobalConfigRepo(t)
 		mGlobalConfigRepo.EXPECT().Get(testCtx).Return(
 			regConfig.CreateGlobalConfig(map[regConfig.Key]regConfig.Value{