+++ migration/config/importer.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/cloudogu/ces-commons-lib/dogu"
+	"github.com/cloudogu/ces-importer/configuration"
 	"github.com/cloudogu/ces-importer/migration"
 	regConfig "github.com/cloudogu/k8s-registry-lib/config"
 	"path"
@@ -23,6 +24,7 @@ type globalConfigRepo interface {
 
 type doguConfigRepo interface {
 	Create(ctx context.Context, doguConfig regConfig.DoguConfig) (regConfig.DoguConfig, error)
+	Get(ctx context.Context, name dogu.SimpleName) (regConfig.DoguConfig, error)
 	SaveOrMerge(ctx context.Context, doguConfig regConfig.DoguConfig) (regConfig.DoguConfig, error)
 	Delete(ctx context.Context, name dogu.SimpleName) error
 }
@@ -46,9 +48,28 @@ type ConfigImporter struct {
 	backupScheduleImporter backupScheduleImporter
 }
 
-func NewConfigImporter(dataBasePath string, configGetter configGetter, globalConfigRepo globalConfigRepo, doguConfigRepo doguConfigRepo, sensitiveDoguConfigRepo doguConfigRepo, backupScheduleClient backupScheduleClient) *ConfigImporter {
-	gci := &cesGlobalConfigImporter{globalConfigRepo}
-	dci := &cesDoguConfigImporter{dataBasePath, doguConfigRepo, sensitiveDoguConfigRepo}
+type ConfigRepos struct {
+	globalConfigRepo        globalConfigRepo
+	doguConfigRepo          doguConfigRepo
+	sensitiveDoguConfigRepo doguConfigRepo
+}
+
+func NewConfigRepos(globalConfigRepo globalConfigRepo, doguConfigRepo doguConfigRepo, sensitiveDoguConfigRepo doguConfigRepo) *ConfigRepos {
+	return &ConfigRepos{
+		globalConfigRepo:        globalConfigRepo,
+		doguConfigRepo:          doguConfigRepo,
+		sensitiveDoguConfigRepo: sensitiveDoguConfigRepo,
+	}
+}
+
+func NewConfigImporter(dataBasePath string, configGetter configGetter, repos ConfigRepos, backupScheduleClient backupScheduleClient, excludedGlobalConfigKeys []string, excludedDoguConfiguration []configuration.DoguConfigurationKeys) *ConfigImporter {
+	excludedConfigKeysByDogu := make(map[string][]string)
+	for _, configs := range excludedDoguConfiguration {
+		excludedConfigKeysByDogu[configs.DoguName] = append(excludedConfigKeysByDogu[configs.DoguName], configs.Keys...)
+	}
+
+	gci := &cesGlobalConfigImporter{repos.globalConfigRepo, excludedGlobalConfigKeys}
+	dci := &cesDoguConfigImporter{dataBasePath, repos.doguConfigRepo, repos.sensitiveDoguConfigRepo, excludedConfigKeysByDogu}
 	bsi := &cesBackupScheduleImporter{backupScheduleClient: backupScheduleClient}
 
 	return &ConfigImporter{