+++ systeminfo/validator.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/cloudogu/ces-importer/migration"
 	"log/slog"
 	"slices"
+
+	doguCommons "github.com/cloudogu/ces-commons-lib/dogu"
+	"github.com/cloudogu/ces-importer/migration"
 )
 
 const (
@@ -55,13 +57,19 @@ func (v *Validator) validateDogus(imInfo *migration.SystemInfo, exInfo *migratio
 	// Create a map of importing dogus for quick lookup
 	imDoguMap := make(map[string]migration.Dogu)
 	for _, d := range imInfo.Dogus {
-		imDoguMap[d.Name] = d
+		imDoguMap[getDoguNameWithoutNamespace(d.Name)] = d
 	}
+
+	// Dogu names in excludedDogus may be names with or without namespace information
+	excludedDoguNames := getExcludedDoguNames(v.excludedDogus)
 
 	// Validate each exporting dogu
 	for _, exDogu := range exInfo.Dogus {
+		doguName := getDoguNameWithoutNamespace(exDogu.Name)
+
 		// Skip excluded dogus
-		if slices.Contains(v.excludedDogus, exDogu.Name) {
+		if slices.Contains(excludedDoguNames, doguName) {
+			slog.Info(fmt.Sprintf("skipping validation for excluded dogu %s", doguName))
 			continue
 		}
 
@@ -74,6 +82,11 @@ func (v *Validator) validateDogus(imInfo *migration.SystemInfo, exInfo *migratio
 		result = validateRegularDogu(exDogu, imDoguMap, result)
 	}
 
+	// remove excluded dogus from map so they don't get flagged as missing on the exporting system
+	for _, excludedDoguName := range excludedDoguNames {
+		delete(imDoguMap, excludedDoguName)
+	}
+
 	// Check for extra dogus in the importing system
 	for name := range imDoguMap {
 		result = errors.Join(result, fmt.Errorf("dogu %s is installed in the importing system but not present in the exporting system \n", name))
@@ -82,20 +95,40 @@ func (v *Validator) validateDogus(imInfo *migration.SystemInfo, exInfo *migratio
 	return result
 }
 
+// getDoguNameWithoutNamespace gets the simple name of a dogu without the namespace
+func getDoguNameWithoutNamespace(doguName string) string {
+	qualifiedDoguName, err := doguCommons.QualifiedNameFromString(doguName)
+	// fall back on name that was passed in
+	if err == nil {
+		doguName = qualifiedDoguName.SimpleName.String()
+	}
+	return doguName
+}
+
+// getExcludedDoguNames gets the list of excluded dogu names without the namespace
+func getExcludedDoguNames(excludedDogus []string) []string {
+	excludedDoguNames := make([]string, len(excludedDogus))
+	for i, doguName := range excludedDogus {
+		excludedDoguNames[i] = getDoguNameWithoutNamespace(doguName)
+	}
+	return excludedDoguNames
+}
+
 // validateRegularDogu validates a single non-nginx dogu and removes it from the map if valid
 func validateRegularDogu(exDogu migration.Dogu, imDoguMap map[string]migration.Dogu, result error) error {
-	imDogu, exists := imDoguMap[exDogu.Name]
+	exDoguName := getDoguNameWithoutNamespace(exDogu.Name)
+	imDogu, exists := imDoguMap[exDoguName]
 	if !exists {
 		return errors.Join(result, fmt.Errorf("dogu %s is not installed (required version: %s) \n", exDogu.Name, exDogu.Version))
 	}
 
 	// Validate version
 	if imDogu.Version != exDogu.Version {
-		result = errors.Join(result, fmt.Errorf("dogu %s is installed in version %s but needs to have version %s \n", exDogu.Name, imDogu.Version, exDogu.Version))
+		result = errors.Join(result, fmt.Errorf("version discrepancy for dogu %s. Source instance version: %s, Target instance version: %s", exDogu.Name, exDogu.Version, imDogu.Version))
 	}
 
 	// Remove validated dogu from map
-	delete(imDoguMap, exDogu.Name)
+	delete(imDoguMap, exDoguName)
 	return result
 }
 
@@ -114,7 +147,7 @@ func validateComponents(imInfo *migration.SystemInfo, exInfo *migration.SystemIn
 
 		// validate that the version is correct
 		if !(imComponent.Version == c.Version) {
-			result = errors.Join(result, fmt.Errorf("component %s is installed in version %s but needs to have version %s \n", c.Name, imComponent.Version, c.Version))
+			result = errors.Join(result, fmt.Errorf("version discrepancy for component %s. Source instance version: %s, Target instance version: %s \n", c.Name, c.Version, imComponent.Version))
 		}
 	}
 	return result